Approximation of the hadron cascade energy density distribution in water as a superposition of cylinders with a uniform distribution density. Test for comparing the results of longitudinal and transverse integration of approximate and exact energy densities distributions

+++ Common/CommonDefinitions.h
@@ -2,7 +2,7 @@
 
 #include <complex>
 
-typedef double fp;
+typedef float fp;
 
 template <class TComplexContainer, class TRealContainer>
 inline void CA2DATA(const TComplexContainer &container, TRealContainer &re_buffer, TRealContainer &im_buffer)
